Add Error 404 Page

+++ src/main/java/backend/repository/IRoleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IRoleRepository extends CrudRepository<Role, Byte> {
+public interface IRoleRepository extends CrudRepository<Role, Integer> {
 }
